Fixed the scroll for the UI
